nycrashes/pyproject.toml
[project]
name = "nycrashes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aws-cdk-lib==2.215.0",
    "constructs>=10.0.0,<11.0.0",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]

[tool.uv.workspace]
members = [
    "backend/populator",
]
